modules/core/app/requirements.txt
streamlit==1.44.0
databricks-sdk==0.50.0
databricks-sql-connector==4.0.2
mlflow==2.22.0
mlflow[databricks]
py3Dmol==2.4.0
biopython
parasail
plotly
openai
rdkit
gseapy
scipy

lib/genesis_workbench-0.1.0-py3-none-any.whl
